requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.29.7
azure-eventhub==5.11.5
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-functions==1.18.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
frozenlist==1.4.1
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
py4j==0.10.9.7
pycparser==2.21
PyJWT==2.8.0
pyspark==3.5.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
python-dotenv==0.1.0